refactor: devided 'parse.c' into some source files
